fix(scrollbar): 64-bit position mapping so huge-range bars map full travel (Int32 overflow at ~2M range on tall viewports)
Two 32-bit products overflowed for large ranges (>= ~1.8M px content) on tall
tracks (>= ~1300 px):
  TyScrollThumbRect:  Offset := (Pos0 * FreeSpace) div Travel
  DragThumbTo:        NewPos := FMin + ((NewTop - TrackStart) * Travel) div FreeSpace
Both widen the dominant factor to Int64 before multiplying.  ThumbLen :=
(APageSize * TrackLen) div Span is safe (max product ~494k, well below 2^31).

Regression tests added to TTyScrollBarHugeRangeTest (pure arithmetic path,
headless): thumb-at-Max lands at track bottom; drag-to-bottom yields Max position.

+++ source/tyControls.ScrollBar.pas
@@ -127,7 +127,7 @@ begin
   if Travel <= 0 then
     Offset := 0
   else
-    Offset := (Pos0 * FreeSpace) div Travel;
+    Offset := Integer((Int64(Pos0) * FreeSpace) div Travel);
   if AKind = sbVertical then
     Result := Rect(ATrack.Left, ATrack.Top + Offset,
       ATrack.Right, ATrack.Top + Offset + ThumbLen)
@@ -510,7 +510,7 @@ begin
   if NewTop < TrackStart then NewTop := TrackStart;
   if NewTop > TrackStart + FreeSpace then NewTop := TrackStart + FreeSpace;
   Travel := FMax - FMin;
-  NewPos := FMin + ((NewTop - TrackStart) * Travel) div FreeSpace;
+  NewPos := FMin + Integer((Int64(NewTop - TrackStart) * Travel) div FreeSpace);
   // Live drag tracking fires scTrack with the proposed value (handler may
   // adjust it); commit through the Position setter (clamps + OnChange).
   if NewPos < FMin then NewPos := FMin;
